- Code clean up refactor - Add Index object - Add Node object
