+++ dask-elk/client.py
@@ -7,166 +7,14 @@ from elasticsearch.helpers import scan, bulk
 from dask.dataframe.utils import make_meta
 import dask.dataframe as dd
 from dask import delayed
-
 import numpy as np
-import pandas as pd
+
+
+from delayed_methods import _elasticsearch_scan
+
 
 class IndexNotFoundException(Exception):
     pass
-
-class DynamicWriteFieldParser(object):
-    compile_re = re.compile('\{(.*?)\}')
-
-    @classmethod
-    def parse_index(cls, record, index):
-        result = index
-        fields = cls.compile_re.findall(index)
-        field_values = {}
-        for field_name_and_frmt in fields:
-            field_name, frmt = field_name_and_frmt.split('|')
-            print field_name, frmt
-            if frmt:
-                field_values[field_name_and_frmt] = record[
-                    field_name].strftime(frmt)
-            else:
-                field_values[field_name_and_frmt] = record[field_name]
-
-        if field_values:
-            result = index.format(**field_values)
-        return result
-
-    @classmethod
-    def parse_doc_type(cls, record, doc_type):
-        result = doc_type
-        fields = cls.compile_re.findall(doc_type)
-        field_values = {}
-        for field_name_and_frmt in fields:
-            field_name, frmt = field_name_and_frmt.split('|')
-            if frmt:
-                field_values[field_name] = record[field_name].strftime(frmt)
-            else:
-                field_values[field_name] = record[field_name]
-        if field_values:
-            result = doc_type.format(**field_values)
-        return result
-
-
-class DocumentParser(object):
-
-    @classmethod
-    def parse_documents(self, documents, meta):
-        """
-
-        :param pandas.DataFrame documents:
-        :param pandas.DataFrame meta:
-        :return: the document formated
-        """
-        meta_columns = meta.columns
-        documents_columns = documents.columns
-
-        missing_columns = self.find_missing_columns(meta,
-                                                    documents)
-        extra_columns = [column for column in documents_columns if
-                         column not in meta_columns]
-
-        for column_name, dtype in missing_columns.iteritems():
-            documents[column_name] = None
-            documents[column_name] = documents[column_name].astype(dtype)
-
-        columns_to_cast = [column_name for column_name in meta_columns if
-                           column_name not in missing_columns]
-
-        for column_name in columns_to_cast:
-            documents[column_name] = documents[column_name].astype(
-                meta.dtypes[column_name])
-
-        if extra_columns:
-            documents = documents.drop(labels=extra_columns, axis=1)
-
-        return documents
-
-    @staticmethod
-    def find_missing_columns(meta, documents):
-        """
-        :param pandas.DataFrame meta:
-        :param pandas.DataFrame documents:
-        :return:
-        :rtype: dict[str, np.dtypes]
-        """
-        meta_columns = meta.columns
-        documents_columns = documents.columns
-
-        diff = [column_name for column_name in meta_columns if
-                column_name not in documents_columns]
-
-        missing_columns = {}
-        for column_name in diff:
-            missing_columns[column_name] = meta.dtypes[column_name]
-
-        return missing_columns
-
-
-def _elasticsearch_scan(client_cls, client_kwargs, meta=None, **params):
-    # This method is executed in the worker's process and here we instantiate
-    # the ES client as it cannot be serialized.
-    # If no data are returned back an empty dataframe is returned with the same
-    # fields and dtypes as meta.
-    client = client_cls(**(client_kwargs or {}))
-    results = []
-
-    for hit in scan(client, **params):
-        result = hit['_source']
-        result['_id'] = hit['_id']
-
-        results.append(result)
-    data_frame = meta
-    if results:
-        data_frame = pd.DataFrame(results)
-        data_frame = DocumentParser.parse_documents(data_frame, meta)
-        data_frame = data_frame[meta.columns]
-    return data_frame
-
-
-def bulk_save(dataframe, client_cls, client_args, **kwargs):
-    """
-
-    :param pandas.DataFrame dataframe:
-    :param client_cls:
-    :param dict[str, T] client_args:
-    :param dict[str, T] kwargs:
-    :return:
-    """
-    elk_client = client_cls(**(client_args or {}))
-    records = dataframe.to_dict(orient='records')
-    actions = []
-
-    for record in records:
-        index = kwargs.get('index').format(**record)
-        doc_type = kwargs.get('doc_type').format(**record)
-        op_type = kwargs['action']
-        for key, value in record.iteritems():
-            if isinstance(value, np.ndarray):
-                record[key] = list(record[key])
-                continue
-
-        action = dict(_index=index, _type=doc_type, _op_type=op_type)
-        if op_type == 'update':
-            action['_id'] = record['_id']
-            del record['_id']
-            action.update({'doc': record})
-        else:
-            action.update(record)
-        actions.append(action)
-
-    print('No of actions to perform: {}'.format(len(actions)))
-    try:
-        print(bulk(elk_client, actions, stats_only=False))
-    except Exception:
-        print actions
-        import traceback
-        traceback.print_exc()
-    finally:
-        return dataframe
 
 
 class DaskElasticClient(object):
