+++ dask-elk/elk_entities/index.py
@@ -1,0 +1,4 @@
+class Index(object):
+    def __init__(self, name, shards):
+        self._name = name
+        self.__shards = shards
